refactor: hook, apis 분리
