+++ src/pages/Portfolio/NewPortfolio.tsx
@@ -7,33 +7,15 @@ import CheckboxSelect from '@/components/common/CheckboxSelect'
 import Textarea from '@/components/common/Textarea'
 import FileUploader from '@/components/common/FileUploader'
 import ImageUploader from '@/components/common/ImageUploader'
-import supabase from '@/lib/supabaseClient'
-import throttle from 'lodash.throttle'
+import type { PortfolioData } from '@/types/portfolio'
+import { useSavePortfolio } from '@/hooks/portfolio/useSavePortfolio'
+import { useCheckValidation } from '@/hooks/portfolio/useCheckValidation'
+import { useUserInfo } from '@/hooks/portfolio/useUserInfo'
+import { useSaveTempPortfolio } from '@/hooks/portfolio/useSaveTempPortfolio'
+import { usePortfolioForm } from '@/hooks/portfolio/usePortfolioForm'
+import { useStickyMenu } from '@/hooks/portfolio/useStickyMenu'
 
-interface UserInfo {
-  id: string
-  email: string
-}
-
-interface PortfolioData {
-  id: string
-  userId: string
-  profileImage: string
-  name: string
-  birthDate: string
-  phone: string
-  email: string
-  title: string
-  content: string
-  career: string
-  interest: string
-  techStack: string[]
-  linkUrl: string
-  fileList: { name: string; url: string }[]
-  viewCount?: number
-  likeCount?: number
-}
-
+// TODOS : 기본정보 마이페이지에서 불러와야함
 const TempData: PortfolioData = {
   id: '',
   userId: '',
@@ -52,8 +34,6 @@ const TempData: PortfolioData = {
   viewCount: 0,
   likeCount: 0
 }
-
-type ValidationError = Partial<Record<keyof PortfolioData, string>>
 
 export const NewPortfolio = () => {
   /* --- 지원분야 라디오 그룹 상태 및 옵션 --- */
@@ -75,31 +55,14 @@ export const NewPortfolio = () => {
     { label: '경력', value: 'senior' }
   ]
 
-  const [userInfo, setUserInfo] = useState<UserInfo | null>(null)
   const [portfolioData, setPortfolioData] = useState<PortfolioData>(TempData)
-  const [errors, setErrors] = useState<ValidationError>({})
+
   /* --- error 체크 --- */
-  const validate = () => {
-    const newErrors: ValidationError = {}
-
-    if (!portfolioData.title.trim()) newErrors.title = '제목을 입력해주세요.'
-    if (!portfolioData.career.trim())
-      newErrors.career = '경력수준을 선택해주세요.'
-    if (!portfolioData.interest.trim())
-      newErrors.interest = '지원분야를 선택해주세요.'
-    if (portfolioData.techStack.length === 0)
-      newErrors.techStack = '기술스택을 선택해주세요.'
-    if (!portfolioData.content.trim())
-      newErrors.content = '소개를 입력해주세요.'
-    if (!portfolioData.linkUrl.trim() && portfolioData.fileList.length === 0)
-      newErrors.linkUrl = 'URL 또는 파일 첨부 중 하나는 반드시 입력해주세요.'
-
-    setErrors(newErrors)
-
-    return Object.keys(newErrors).length == 0
-  }
+  const { validate, errors, setErrors } = useCheckValidation()
 
   /* --- 로그인 시 유저정보 불러오기 --- */
+  const { userInfo } = useUserInfo()
+
   useEffect(() => {
     if (userInfo?.email) {
       setPortfolioData(prev => ({
@@ -109,111 +72,20 @@ export const NewPortfolio = () => {
     }
   }, [userInfo])
 
-  useEffect(() => {
-    const fetchUserInfo = async () => {
-      const {
-        data: { user },
-        error
-      } = await supabase.auth.getUser()
-
-      if (error) {
-        console.error('유저 정보 불러오기 실패:', error)
-        return
-      }
-
-      if (user && user.email) {
-        setUserInfo({ id: user.id, email: user.email })
-      }
-    }
-
-    fetchUserInfo()
-  }, [])
-
   /* --- 입력값 변경 시 상태 저장 및 관련 에러 제거 --- */
-  const handleChangeForm = <K extends keyof PortfolioData>(
-    key: K,
-    value: PortfolioData[K]
-  ) => {
-    setPortfolioData(prev => ({
-      ...prev,
-      [key]: value
-    }))
-
-    setErrors(prevErrors => {
-      const newErrors = { ...prevErrors }
-
-      if (key === 'linkUrl' || key === 'fileList') {
-        delete newErrors.linkUrl
-        delete newErrors.fileList
-      } else if (newErrors[key]) {
-        delete newErrors[key]
-      }
-      return newErrors
-    })
-  }
+  const { handleChangeForm } = usePortfolioForm(setPortfolioData, setErrors)
 
   /* --- 임시저장 --- */
-  const handleSaveTemp = async () => {
-    try {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const { viewCount, likeCount, ...rest } = portfolioData
-
-      const { error } = await supabase
-        .from('TempPortfolio')
-        .upsert(
-          { ...rest, userId: userInfo?.id, id: undefined },
-          { onConflict: 'id' }
-        )
-      if (error) throw error
-      alert('임시저장되었습니다.')
-    } catch (error) {
-      alert('임시저장이 실패하였습니다. 다시 시도해주세요.')
-      console.error(error)
-    }
-  }
+  const { handleSaveTemp } = useSaveTempPortfolio({ portfolioData, userInfo })
 
   /* --- 저장 --- */
-  const handleSave = async () => {
-    if (!validate()) {
-      alert('빠진 부분이 있는지 확인해주세요.')
-      return
-    }
-    try {
-      const { error } = await supabase.from('Portfolio').insert({
-        ...portfolioData,
-        userId: userInfo?.id,
-        id: undefined,
-        viewCount: 0,
-        likeCount: 0
-      })
-      if (error) throw error
-      alert('포트폴리오가 저장되었습니다.')
-    } catch (error) {
-      alert('저장에 실패하였습니다. 다시 시도해주세요.')
-      console.error(error)
-    }
-  }
+  const { handleSave } = useSavePortfolio({ portfolioData, userInfo, validate })
 
   /* --- 타이틀 및 버튼 sticky --- */
   const stickyRef = useRef<HTMLDivElement | null>(null)
   const [isSticky, setIsSticky] = useState(false)
 
-  useEffect(() => {
-    const handleScroll = throttle(() => {
-      if (!stickyRef.current) return
-
-      const stickyPosition = stickyRef.current.getBoundingClientRect().top ?? 0
-      const yes = stickyPosition <= 0
-      setIsSticky(yes)
-    }, 200)
-
-    window.addEventListener('scroll', handleScroll)
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll)
-      handleScroll.cancel()
-    }
-  }, [])
+  useStickyMenu(stickyRef, setIsSticky)
 
   return (
     <div className={S.container}>
